# openapi/vsdd-sprint-tracker.openapi.yaml
openapi: 3.1.0
info:
  title: VSDD Sprint Tracker API
  version: 1.0.0
  description: >-
    REST/JSON contract for the VSDD Sprint Tracker (design.md §6). Component
    schemas mirror the implemented frontend TypeScript domain types and Zod
    schemas (src/domain/*, src/api/repository.ts) so the server and the
    replaceable mock repository share one contract.


    Persistence is document-oriented and vendor-neutral (design.md §4a): each
    team/checkpoint update is a single versioned JSON document with a stable
    query envelope plus a flexible, schema-versioned payload. Optimistic
    concurrency uses `revision` as an ETag; a stale revision on write returns
    409 DRAFT_REVISION_CONFLICT and overwrites nothing.


    This document defines the contract only — no server logic is implied
    (Phase 7.3+ implements the endpoints).
  contact:
    name: VSDD Sprint Tracker
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /api/v1
    description: Version 1 API base path

# Every request is authenticated and authorised server-side against
# programme/team assignments (design.md §13). Applied globally; individual
# operations may narrow scope in later phases.
security:
  - bearerAuth: []
...
paths:
...
  /updates/{versionId}/decisions:
    post:
      operationId: recordDecision
      tags: [leadership]
      summary: Record a leadership decision against a submitted version
      parameters:
        - $ref: '#/components/parameters/VersionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DecisionRequest'
      responses:
        '201':
          description: The recorded leadership decision
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeadershipDecision'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    get:
      operationId: getDecisions
      tags: [leadership]
      summary: Leadership decisions recorded against a version
      parameters:
        - $ref: '#/components/parameters/VersionId'
      responses:
        '200':
          description: Decisions for the version
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LeadershipDecision'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        Enterprise OIDC access token. The application receives a stable subject
        id and group/role claims (design.md §2, §13). An approved same-site
        secure-cookie session is an equivalent accepted pattern.
  parameters:
...
    VersionId:
      name: versionId
      in: path
      required: true
      schema:
        type: string
...
  responses:
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: Authenticated but not authorised for this programme/team
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: PERMISSION_DENIED
              message: You do not have permission to edit this team update.
              correlationId: 7b1c0f2a-2a3e-4c9a-9e0b-2a1f3c4d5e6f
              fieldErrors: []
    NotFound:
      description: The requested resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: NOT_FOUND
              message: The requested update could not be found.
              correlationId: 7b1c0f2a-2a3e-4c9a-9e0b-2a1f3c4d5e6f
              fieldErrors: []
    ValidationError:
      description: The request failed validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: VALIDATION_FAILED
              message: Fix the highlighted fields and try again.
              correlationId: 7b1c0f2a-2a3e-4c9a-9e0b-2a1f3c4d5e6f
              fieldErrors:
                - path: goals.business
                  message: This field is required.
...
  schemas:
...
    LeadershipDecision:
      type: object
      additionalProperties: false
      required: [id, updateVersionId, decision, ownerSubject, status, createdAt]
      properties:
        id:
          type: string
        updateVersionId:
          type: string
        decision:
          type: string
        ownerSubject:
          type: string
        dueDate:
          type: string
          format: date
        status:
          type: string
          enum: [OPEN, CLOSED]
        createdAt:
          type: string
          format: date-time

    # ---------------------------------------------------------------------
    # Version comparison (server/src/domain/versionComparison.ts, R14.3)
    # ---------------------------------------------------------------------
...
    DecisionRequest:
      type: object
      additionalProperties: false
      required: [decision]
      properties:
        decision:
          type: string
          minLength: 1
        dueDate:
          type: string
          format: date
...
    ErrorCode:
      type: string
      enum:
        - DRAFT_REVISION_CONFLICT
        - PERMISSION_DENIED
        - WINDOW_CLOSED
        - ALREADY_SUBMITTED
        - INVALID_STATE
        - NOT_FOUND
        - SAVE_FAILED
        - VALIDATION_FAILED
    FieldError:
      type: object
      additionalProperties: false
      required: [path, message]
      properties:
        path:
          type: string
          description: Dot path, e.g. `goals.business` or `exceptions.0.owner`.
        message:
          type: string
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, correlationId, fieldErrors]
          properties:
            code:
              $ref: '#/components/schemas/ErrorCode'
            message:
              type: string
              description: User-facing copy explaining what happened and the next action.
            correlationId:
              type: string
            fieldErrors:
              type: array
              items:
                $ref: '#/components/schemas/FieldError'
...